move terminal into os dir
